Refactor a little and go forward on MessageManager.
In this commit, I transformed Peer into PeerImpl and I added a new
Interface Peer. In that way, I just have to implements Peer in the
way I want in the test to test specifically what I want.

I implemented the methods of MessageManager and I tested it with
the JUnit test TestMessageManager.

I added the Callback TomDeliveredCallback. It is equivalement to
the previous DeliveredCallback, but it is specific to the layer
totally ordered multicast.

Perhaps multicast already work ^^ But I haven't tested it yet.
I just made JUnit tests. I would need at least an extra hour.