Solve the problem by correcting the burst mode :P
I fact, you send each time to the manager a string representing the
InetSocketAddress. However, (I don't know why) but the inetsocketAddress
is not displayed in the same way on a peer and on the others...
The at the end, the manager concluded that the messages were different
since the displayed address was different, but it was the same!