Fix an issue with message layer refactoring
The new message layer was not correctly handling the case of using the loopback address for every peer of the group. This case should be handled apart from the reset of the code (see ByteUtil.readInetSocketAddress).
This is due to the underlying network stack and the very definition of what Loopback is.