Update PeerImpl
When a new member connects to a peer (ie accepted callback is triggered) we no longer add it to the group. The new member is instead added to waitingMembers. That means the burstMode and the basic tests of tom no longer works, we have to finish the implementation of member joining to get it work again.